backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2
Pillow==10.0.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8